+++ crates/config/src/project/mod.rs
@@ -1,5 +1,6 @@
 mod dep_config;
 mod docker_config;
+mod language;
 mod overrides_config;
 mod owners_config;
 mod task_config;
@@ -7,6 +8,7 @@ mod task_options_config;
 
 pub use dep_config::*;
 pub use docker_config::*;
+pub use language::*;
 pub use overrides_config::*;
 pub use owners_config::*;
 pub use task_config::*;